virt_mshv_vtl: Move some cvm-only interrupt handling code into cvm-only locations.
